Improve components structure, take utils from Perfomance component to separate files
