+++ src/app/utils/formatAsUsd.ts
@@ -1,0 +1,6 @@
+export const formatAsUSD = (value: number) => {
+    return new Intl.NumberFormat('en-US', {
+      style: 'currency',
+      currency: 'USD',
+    }).format(value);
+};
