+++ src/app/utils/formatDate.ts
@@ -1,0 +1,7 @@
+export const formatDate = (date: Date) => {
+    return new Intl.DateTimeFormat('en-US', {
+        year: 'numeric',
+        month: 'short',
+        day: '2-digit',
+    }).format(new Date(date));
+};
